+++ Assets/Warlander/Deedplanner/Bridges/BridgeMaterialChangeCommand.cs
@@ -26,13 +26,11 @@ namespace Warlander.Deedplanner.Bridges
         public void Execute()
         {
             _bridge.Rebuild(_map, _newMaterial, _newSegments, _bridge.AdditionalData);
-            _map.RefreshBridgesRendering();
         }
 
         public void Undo()
         {
             _bridge.Rebuild(_map, _oldMaterial, _oldSegments, _bridge.AdditionalData);
-            _map.RefreshBridgesRendering();
         }
 
         public void DisposeUndo()