+++ Assets/Warlander/Deedplanner/Bridges/BridgeSegmentsChangeCommand.cs
@@ -21,13 +21,11 @@ namespace Warlander.Deedplanner.Bridges
         public void Execute()
         {
             _bridge.Rebuild(_map, _bridge.Data, _newSegments, _bridge.AdditionalData);
-            _map.RefreshBridgesRendering();
         }
 
         public void Undo()
         {
             _bridge.Rebuild(_map, _bridge.Data, _oldSegments, _bridge.AdditionalData);
-            _map.RefreshBridgesRendering();
         }
 
         public void DisposeUndo()